requirements.txt
python-dotenv
langfuse
google-cloud-aiplatform
google.genai